+++ pica-matcher/src/field_matcher.rs
@@ -579,7 +579,9 @@ fn parse_field_matcher_composite(
 }
 
 /// Parse a field matcher expression.
-fn parse_field_matcher(i: &[u8]) -> ParseResult<FieldMatcher> {
+pub(crate) fn parse_field_matcher(
+    i: &[u8],
+) -> ParseResult<FieldMatcher> {
     alt((
         ws(parse_field_matcher_composite),
         ws(parse_field_matcher_group),